Rename dc2_catalog->dc2_dm_catalog, DC2Catalog->DC2DMCatalog.
Make it clearer that this file and class are primarily targeted at
catalogs that are the result of LSST DM Science Pipelines.
